requirements.txt
aiobotocore==1.0.3
aiofiles==0.5.0
aiohttp==3.6.2
aioitertools==0.6.0
aiosqlite==0.12.0
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
botocore==1.15.32
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.4.14
httptools==0.1.1
httpx==0.9.3
hyperframe==5.2.0
idna==2.9
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.7.5
pycodestyle==2.5.0
pyflakes==2.1.1
PyPika==0.36.3
python-dateutil==2.8.1
rfc3986==1.4.0
sanic==19.12.2
Sanic-Auth==0.2.0
Sanic-CookieSession==0.2.0
sanic-jinja2==0.7.5
Sanic-WTF==0.5.0
six==1.14.0
sniffio==1.1.0
tortoise-orm==0.16.5
typing-extensions==3.7.4.2
urllib3==1.25.8
websockets==8.1
wrapt==1.12.1
WTForms==2.2.1
yarl==1.4.2
gunicorn==20.0.4
